Revert "CI: Only run builds after successful linting"
This reverts commits d19d8b4e3a79fd441c8fa7265c506c84af3d6e97 and
b3f1d71d35c588620331c952d6e7e1ae2ff0f622.

The commit linter needs to run on event `pull_request_target` to have
access to its secret token, which means we cannot have a dependency on
that job from another workflow that is run as a result of the
`pull_request` event.

Additionally, the linters were no longer run for first-time
contributors. This isn't a huge problem but it was nice that a
preliminary check took place before running the full CI on their PRs.